requirements.txt
typedb-client==2.9.0
networkx[default]
tabulate